:ghost: Refine usage of (scm, command) hub packages. (#83)
The implementation for:
- Running commands (cmd)
- SSH (ssh)
- Git (scm)
- Subversion (scm)

Was moved to the hub https://github.com/konveyor/tackle2-hub/pull/932.

As a result, the implementations was refactored and functionality
delegated to the imported hub packages. The primary role now is to
provide integration with the addon environment. Mainly to fetch
resources such as proxies, settings and identities using the binding.
This is done in the `New()` (factory).

This is a refinement of:
https://github.com/konveyor/tackle2-addon/pull/82

<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

* **Refactor**
* Simplified SCM initialization and clarified repository/identity
handling, improving remote and proxy behavior.
* Reworked output writer to use a shared buffer and bytes-based end
reporting, removing legacy I/O surface.

* **Chores**
  * Updated project dependencies to newer versions.

<sub>✏️ Tip: You can customize this high-level summary in your review
settings.</sub>
<!-- end of auto-generated comment: release notes by coderabbit.ai -->

---------

Signed-off-by: Jeff Ortel <jortel@redhat.com>